% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validCategoricalNames}
\alias{validCategoricalNames}
\title{check if a vector of Categorical model name is correct.}
\usage{
validCategoricalNames(names)
}
\arguments{
\item{names}{a vector of character}
}
\description{
check if a vector of Categorical model name is correct.
}
\keyword{internal}

