% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkDiagGaussianNames}
\alias{checkDiagGaussianNames}
\title{check if a vector of diagonal Gaussian model name comply}
\usage{
checkDiagGaussianNames(modelNames)
}
\description{
check if a vector of diagonal Gaussian model name comply
}
\keyword{internal}

