% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ClusterAlgo}
\alias{ClusterAlgo}
\alias{ClusterAlgo-class}
\title{[\code{\linkS4class{ClusterAlgo}}] class for Cluster algorithms.}
\description{
This class encapsulate the parameters of estimation algorithms of the rtkpp Cluster
methods.
}
\section{Slots}{

\describe{
\item{\code{algo}}{A character string with the algorithm. Possible values: "SEM", "CEM", "EM", "SemiSEM". Default value: "EM".}

\item{\code{nbIteration}}{Integer defining the maximal number of iterations. Default value: 200.}

\item{\code{epsilon}}{real defining the epsilon value for the algorithm. epsilon is note used if \code{algo} is sem.
Default value: 1e-07.}
}}
\examples{
getSlots("ClusterAlgo")
new("ClusterAlgo")
new("ClusterAlgo", algo="SEM", nbIteration=1000)
}

