% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clusterCategorical}
\alias{clusterCategorical}
\title{Create an instance of the [\code{\linkS4class{ClusterCategorical}}] class}
\usage{
clusterCategorical(data, nbCluster = 2, modelNames = NULL,
  strategy = clusterStrategy(), criterion = "ICL")
}
\arguments{
\item{data}{frame or matrix containing the data. Rows correspond to observations
and columns correspond to variables. If the data set contains NA values, they
will be estimated during the estimation process.}

\item{nbCluster}{[\code{\link{vector}}] listing the number of clusters to test.}

\item{modelNames}{[\code{\link{vector}}] of models names to run. By default
all categorical models are estimated.}

\item{strategy}{a [\code{\linkS4class{ClusterStrategy}}] object containing
the strategy to run. clusterStrategy() method by default.}

\item{criterion}{character defining the criterion to select the best model.
The best model is the one with the lowest criterion value.
Possible values: "BIC", "AIC", "ICL". Default is "ICL".}
}
\value{
An instance of the [\code{\linkS4class{ClusterCategorical}}] class.
}
\description{
This function computes the optimal Categorical mixture model according
to the [\code{criterion}] among the list of model given in [\code{modelNames}]
and the number of clusters given in [\code{nbCluster}], using the strategy
specified in [\code{strategy}].
}
\examples{
## A quantitative example with the birds data set
data(birds)
## add 10 missing values
x = birds;
x[round(runif(5,1,nrow(birds))), 2] <- NA
x[round(runif(5,1,nrow(birds))), 4] <- NA
## with default values
model <- clusterCategorical(data=x, nbCluster=2:3, strategy = clusterFastStrategy())

## use graphics functions
\dontrun{
plot(model)
}

## get summary
summary(model)
}
\author{
Serge Iovleff
}

