% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ClusterDiagGaussian}
\alias{ClusterDiagGaussian}
\alias{ClusterDiagGaussian-class}
\title{Definition of the [\code{\linkS4class{ClusterDiagGaussian}}] class}
\description{
This class defines a diagonal Gaussian mixture Model.
}
\details{
This class inherits from the [\code{\linkS4class{IClusterModel}}] class.
A diagonal gaussian model is a mixture model of the form:
\deqn{
  f({x}|\boldsymbol{\theta})
  =\sum_{k=1}^K p_k \prod_{j=1}^d \phi(x_j;\mu_{jk},\sigma^2_{jk})
   \quad x \in {R}^d.
}
}
\section{Slots}{

\describe{
\item{\code{mean}}{Matrix with the mean of the jth variable in the kth cluster.}

\item{\code{sigma}}{Matrix with the standard deviation of the jth variable in the kth cluster.}
}}
\examples{
getSlots("ClusterDiagGaussian")
data(geyser)
new("ClusterDiagGaussian", data=geyser)
}
\author{
Serge Iovleff
}
\seealso{
[\code{\linkS4class{IClusterModel}}] class
}

