% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusterDiagGaussianModel}
\alias{ClusterDiagGaussianModel}
\alias{ClusterDiagGaussianModel-class}
\title{Definition of the [\code{\linkS4class{ClusterDiagGaussianModel}}] class}
\description{
This class defines a diagonal Gaussian mixture Model.
}
\details{
This class inherits from the [\code{\linkS4class{IClusterModel}}] class.
A diagonal gaussian model is a mixture model of the form:
\deqn{
  f({x}|\boldsymbol{\theta}) \\
  =\sum_{k=1}^K p_k \prod_{j=1}^d \psi(x_j;\mu_{jk},\sigma^2_{jk}) \\
   \quad {x} \in {R}^d.
}
}
\section{Slots}{

\describe{
\item{\code{meankj}}{Matrix with the mean of the jth variable in the kth cluster.}

\item{\code{sigma2kj}}{Matrix with the variance of the jth variable in the kth cluster.}
}}
\examples{
getSlots("ClusterDiagGaussianModel")
  data(geyser)
  new("ClusterDiagGaussianModel", data=geyser)
}
\author{
Serge Iovleff
}

