% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusterGammaModel}
\alias{ClusterGammaModel}
\alias{ClusterGammaModel-class}
\title{Definition of the [\code{\linkS4class{ClusterGammaModel}}] class}
\description{
This class defines a gamma mixture Model. Inherits from the
[\code{\linkS4class{IClusterModel}}] class. A gamma mixture model is
a mixture model of the form
}
\details{
\deqn{
  f({x}|\boldsymbol{\theta}) \\
  =\sum_{k=1}^K p_k \prod_{j=1}^d \gamma(x_j;a_{jk},b_{jk}) \\
   \quad {x} \in {R}^d.
}
}
\section{Slots}{

\describe{
\item{\code{ajk}}{Matrix with the shape of the jth variable in the kth cluster.}

\item{\code{bjk}}{Matrix with the scale of the jth variable in the kth cluster.}
}}
\examples{
getSlots("ClusterGammaModel")
  new("ClusterGammaModel", data=iris[1:4])
}
\author{
Serge Iovleff
}

