% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ClusterHeterogeneous}
\alias{ClusterHeterogeneous}
\alias{ClusterHeterogeneous-class}
\title{Definition of the [\code{\linkS4class{ClusterHeterogeneous}}] class}
\description{
This class defines an Heterogeneous mixture Model.
}
\details{
This class inherits from the [\code{\linkS4class{IClusterModelBase}}] class.
A diagonal gaussian model is a mixture model of the form:
\deqn{
  f({x}|\boldsymbol{\theta})
  =\sum_{k=1}^K p_k \prod_{j=1}^d \phi(x_j;\mu_{jk},\sigma^2_{jk})
   \quad x \in {R}^d.
}
}
\section{Slots}{

\describe{
\item{\code{ldata}}{a list of IClusterComponent.}
}}
\examples{
getSlots("ClusterHeterogeneous")
}
\author{
Serge Iovleff
}
\seealso{
[\code{\linkS4class{IClusterModelBase}}] class
}

