% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusteringAlgo}
\alias{ClusteringAlgo}
\alias{ClusteringAlgo-class}
\title{[\code{\linkS4class{ClusteringAlgo}}] class for clustering algorithms.}
\description{
This class encapsulate the parameters of estimation algorithms of the rtkpp Clustering
methods.
}
\section{Slots}{

\describe{
\item{\code{algo}}{A character string with the algorithm. Possible values: "SEM", "CEM", "EM". Default value: "EM".}

\item{\code{nbIteration}}{Integer defining the maximal number of iterations. Default value: 200.}

\item{\code{epsilon}}{real defining the epsilon value for the algorithm. epsilon is note used if \code{algo} is sem.
Default value: 1e-07.}
}}
\examples{
new("ClusteringAlgo")
new("ClusteringAlgo", algo="SEM", nbIteration=1000)
getSlots("ClusteringAlgo")
}

