% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusteringCategoricalModel-class}
\alias{ClusteringCategoricalModel-class}
\title{Definition of the [\code{\linkS4class{ClusteringCategoricalModel}}] class}
\description{
This class defines a categorical mixture Model. Inherits from the
[\code{\linkS4class{IClusteringModel}}] class. A categorical mixture model is
a mixture model of the form
}
\details{
\deqn{
  f({x}|\boldsymbol{\theta}) \\
  =\sum_{k=1}^K p_k \prod_{j=1}^d \mathcal{M}(x_j;p_{jk},1) \\
   \quad {x} \in \{1,\ldots,L\}^d.
}
}
\section{Slots}{

\describe{
\item{\code{pljk}}{Array with the probability of the jth variable in the kth cluster
to be l.}
}}
\examples{
getSlots("ClusteringCategoricalModel")
  new("ClusteringCategoricalModel", data=iris[1:4])
}
\author{
Serge Iovleff
}

