% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusteringDiagGaussianModel}
\alias{ClusteringDiagGaussianModel}
\alias{ClusteringDiagGaussianModel-class}
\title{Definition of the [\code{\linkS4class{ClusteringDiagGaussianModel}}] class}
\description{
This class defines a diagonal Gaussian mixture Model.
}
\details{
This class inherits from the [\code{\linkS4class{IClusteringModel}}] class.
A diagonal gaussian model is a mixture model of the form:
\deqn{
  f({x}|\boldsymbol{\theta}) \\
  =\sum_{k=1}^K p_k \prod_{j=1}^d \psi(x_j;\mu_{jk},\sigma^2_{jk}) \\
   \quad {x} \in {R}^d.
}
}
\section{Slots}{

\describe{
\item{\code{meanjk}}{Matrix with the mean of the jth variable in the kth cluster.}

\item{\code{sigma2jk}}{Matrix with the variance of the jth variable in the kth cluster.}
}}
\examples{
getSlots("ClusteringDiagGaussianModel")
  new("ClusteringDiagGaussianModel", data=iris[1:4])
}
\author{
Serge Iovleff
}

