% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusteringInit}
\alias{ClusteringInit}
\alias{ClusteringInit-class}
\title{[\code{\linkS4class{ClusteringInit}}] class}
\description{
This class encapsulates the parameters of initialization methods of the
rtkpp clustering estimation method.
}
\section{Slots}{

\describe{
\item{\code{method}}{Character string with the initialization method to use. Default value: "random"}

\item{\code{nbInit}}{Integer defining the number of initialization to perform. Default value: 1.}

\item{\code{algo}}{An instance of \code{\linkS4class{ClusteringAlgo}} class.
Default value: \code{clusteringAlgo("SEM", 20, 0)}.}
}}
\examples{
getSlots("ClusteringInit")
  new("ClusteringInit")
}
\author{
Serge Iovleff
}

