% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClusteringStrategy}
\alias{ClusteringStrategy}
\alias{ClusteringStrategy-class}
\title{Constructor of [\code{\linkS4class{ClusteringStrategy}}] class}
\description{
This class encapsulate the parameters of the strategy of estimation of the rtkpp
Clustering models.
}
\section{Slots}{

\describe{
\item{\code{nbTry}}{Integer defining the number of tries. Default value: 1.}

\item{\code{initMethod}}{A [\code{\linkS4class{ClusteringInit}}] object defining the way to
initialize the estimation method.}

\item{\code{shortAlgo}}{A [\code{\linkS4class{ClusteringAlgo}}] object defining the algorithm
to use during the short runs of the estimation method.}

\item{\code{longAlgo}}{A [\code{\linkS4class{ClusteringAlgo}}] object defining the algorithm
  to use during the long run of the estimation method.}
}}
\examples{
new("ClusteringStrategy")
  new("ClusteringStrategy", shortAlgo=clusteringAlgo("SEM",1000))
  getSlots("ClusteringStrategy")
}
\author{
Serge Iovleff
}

