% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ClusterInit}
\alias{ClusterInit}
\alias{ClusterInit-class}
\title{Constructor of the [\code{\linkS4class{ClusterInit}}] class}
\description{
This class encapsulates the parameters of initialization methods of the
rtkpp Cluster estimation method.
}
\section{Slots}{

\describe{
\item{\code{method}}{Character string with the initialization method to use. Default value: "random"}

\item{\code{nbInitRun}}{Integer defining the number of initialization to perform. Default value: 5.}

\item{\code{algo}}{An instance of \code{\linkS4class{ClusterAlgo}} class.
Default value: \code{clusterAlgo("SEM", 20, 0)}.}
}}
\examples{
getSlots("ClusterInit")
  new("ClusterInit")
  new("ClusterInit", nbInitRun=1)
}
\author{
Serge Iovleff
}

