% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clusterPoisson}
\alias{clusterPoisson}
\title{Create an instance of the [\code{\linkS4class{ClusterPoisson}}] class}
\usage{
clusterPoisson(data, nbCluster = 2, modelNames = clusterPoissonNames(),
  strategy = clusterFastStrategy(), criterion = "ICL")
}
\arguments{
\item{data}{frame or matrix containing the data. Rows correspond to observations
and columns correspond to variables. If the data set contains NA values, they
will be estimated during the estimation process.}

\item{nbCluster}{[\code{\link{vector}}] listing the number of clusters to test.}

\item{modelNames}{[\code{\link{vector}}] of model names to run. By default all
poisson models are estimated.  All the model names are given by
the method [\code{\link{clusterPoissonNames}}].}

\item{strategy}{a [\code{\linkS4class{ClusterStrategy}}] object containing
the strategy to run. clusterStrategy() method by default.}

\item{criterion}{character defining the criterion to select the best model.
The best model is the one with the lowest criterion value.
Possible values: "BIC", "AIC", "ICL". Default is "ICL".}
}
\value{
An instance of the [\code{\linkS4class{ClusterPoisson}}] class.
}
\description{
This function computes the optimal poisson mixture model according
to the [\code{criterion}] among the list of model given in [\code{modelNames}]
and the number of clusters given in [\code{nbCluster}], using the strategy
specified in [\code{strategy}].
}
\examples{
## A quantitative example with the DebTrivedi data set.
data(DebTrivedi)
## use 500 first observation of the Trivedi data set
dt <- DebTrivedi[1:500, c(1, 6,8, 15)]

model <- clusterPoisson(data=dt, nbCluster=2:3, strategy = clusterFastStrategy())

## use graphics functions
\dontrun{
plot(model)
}

## get summary
summary(model)
## print model
print(model)
## get estimated missing values
missingValues(model)
}
\author{
Serge Iovleff
}

