% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clusterPoissonNames}
\alias{clusterPoissonNames}
\alias{clusterValidPoissonNames}
\title{Build a vector of Poisson model names.}
\usage{
clusterPoissonNames(prop = "all", proportional = "both")

clusterValidPoissonNames(names)
}
\arguments{
\item{prop}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{proportional}{A character string equal to "yes", "no" or "both". Default is "both".}

\item{names}{a vector of character}
}
\value{
A vector of character with the model names.
}
\description{
In a poisson model, we can build 4 models:
\enumerate{
 \item {The proportions can be equal or free.}
 \item {The intensities can be proportional or free for all the variables.}
}

check if a vector of Poisson model name is correct.
}
\details{
The model names are summarized in the following array:
\tabular{lll}{
 Model Name      \tab Proportions \tab Proportional  \cr
 poisson_p_ljk   \tab Equal       \tab No            \cr
 poisson_p_ljpk  \tab Equal       \tab Yes           \cr
 poisson_pk_ljk  \tab Free        \tab No            \cr
 poisson_pk_ljpk \tab Free        \tab Yes           \cr
}
}
\examples{
clusterPoissonNames()
clusterPoissonNames("all", "yes") # same as c( "poisson_pk_ljlk", "poisson_p_ljlk")
}
\keyword{internal}

