% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagGaussianNames}
\alias{diagGaussianNames}
\alias{validDiagGaussianNames}
\title{diagGaussianNames: get the names of the diagonal Gaussian models}
\usage{
diagGaussianNames(prop = "all", varianceVariables = "all",
  varianceClusters = "all")

validDiagGaussianNames(modelNames)
}
\arguments{
\item{prop}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{varianceVariables}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{varianceClusters}{A character string equal to "equal", "free" or "all". Default is "all".}
}
\description{
Build a vector of diagonal Gaussian model names.

check if a vector of diagonal Gaussian model name comply
}
\details{
In a diagonal Gaussian model, we assume that the variance
matrices are diagonal in each cluster. This gives rise to 8 models:
\enumerate{
 \item {The proportions can be equal or free.}
 \item {The variances can be equal or free for all the variables.}
 \item {The variances can be equal or free for all the clusters.}
}

The model names are summarized in the following array:
\tabular{llll}{
    Model Name      \tab Prop. \tab Variance of the variables \tab Variance of the clusters \cr
    gaussian_p_sjk  \tab Equal \tab Free  \tab Free  \cr
    gaussian_p_sj   \tab Equal \tab Free  \tab Equal \cr
    gaussian_p_sk   \tab Equal \tab Equal \tab Free  \cr
    gaussian_p_s    \tab Equal \tab Equal \tab Equal \cr
    gaussian_pk_sjk \tab Free  \tab Free  \tab Free  \cr
    gaussian_pk_sj  \tab Free  \tab Free  \tab Equal \cr
    gaussian_pk_sk  \tab Free  \tab Equal \tab Free  \cr
    gaussian_pk_s   \tab Free  \tab Equal \tab Equal \cr
}
}
\examples{
diagGaussianNames()
diagGaussianNames("all", "equal", "free") # same as c("gaussian_p_sk", "gaussian_pk_sk")
}
\keyword{internal}

