% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{IClusteringModelNames}
\alias{IClusteringModelNames}
\alias{IClusteringModelNames-class}
\title{Interface base class [\code{\linkS4class{IClusteringModelNames}}]}
\description{
This class is a virtual class of the Clustering model names.
}
\section{Slots}{

\describe{
\item{\code{listModels}}{character list of models.}

\item{\code{typeModels}}{character string with the type of model selected. Possible values:
"equal", "free", all". Default value: "all".}
}}
\examples{
getSlots("IClusteringModelNames")
}
\author{
Serge Iovleff
}

