% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{IClusterModel}
\alias{IClusterModel}
\alias{IClusterModel-class}
\title{Interface Class [\code{\linkS4class{IClusterModel}}] for Cluster models.}
\description{
This class encapsulate the common parameters of all the Cluster models.
}
\details{
A Cluster model is a model of the form
\deqn{
  f({x}|\boldsymbol{\theta}) \\
    \sum_{k=1}^K p_k f({x};\boldsymbol{\lambda}_k,\boldsymbol{\alpha}) \\
   \quad {x} \in J.
}
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{\link{matrix}} with the data set to cluster.}

\item{\code{nbCluster}}{Integer with the number of cluster of the model.}

\item{\code{pk}}{Vector with the proportions of each mixture.}

\item{\code{tik}}{Matrix with the posterior probability of the ith individual to belong to kth cluster.}

\item{\code{fi}}{Vector with the mixture probabilities of the individuals.}

\item{\code{zi}}{Vector with the class label of the individuals.}

\item{\code{lnLikelihood}}{Real given the ln-liklihood of the Cluster model.}

\item{\code{criterion}}{Real given the value of the AIC or BIC criterion.}

\item{\code{modelName}}{mixture model name.}
}}
\examples{
getSlots("IClusterModel")
}
\author{
Serge Iovleff
}

