% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusterAlgo-method}
\alias{initialize,ClusterAlgo-method}
\title{Initialize an instance of the [\code{\linkS4class{ClusterAlgo}}] class.}
\usage{
\S4method{initialize}{ClusterAlgo}(.Object, algo, nbIteration, epsilon)
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

