% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusterDiagGaussianModel-method}
\alias{initialize,ClusterDiagGaussianModel-method}
\title{Initialize an instance of the [\code{\linkS4class{ClusterDiagGaussianModel}}] class.}
\usage{
\S4method{initialize}{ClusterDiagGaussianModel}(.Object, data, nbCluster = 2,
  modelName = "gaussian_sjk")
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

