% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusterGammaModel-method}
\alias{initialize,ClusterGammaModel-method}
\title{Initialize an instance of the [\code{\linkS4class{ClusterGammaModel}}] class.}
\usage{
\S4method{initialize}{ClusterGammaModel}(.Object, data, nbCluster = 2,
  modelName = "gamma_ajk_bjk", hasFreeProportions = TRUE)
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

