% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusteringDiagGaussianModel-method}
\alias{initialize,ClusteringDiagGaussianModel-method}
\title{Initialize an instance of the [\code{\linkS4class{ClusteringDiagGaussianModel}}] class.}
\usage{
\S4method{initialize}{ClusteringDiagGaussianModel}(.Object, data,
  nbCluster = 2, modelName = "gaussian_sjk", hasFreeProportions = TRUE)
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

