% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusteringInit-method}
\alias{initialize,ClusteringInit-method}
\title{Create an instance of the [\code{\linkS4class{ClusteringInit}}] class using new/initialize.}
\usage{
\S4method{initialize}{ClusteringInit}(.Object, method = "random",
  nbInit = 5, algo = clusteringAlgo("SEM", 20, 0))
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

