% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusteringStrategy-method}
\alias{initialize,ClusteringStrategy-method}
\title{Create an instance of the [\code{\linkS4class{ClusteringStrategy}}] class using new/initialize.}
\usage{
\S4method{initialize}{ClusteringStrategy}(.Object, nbTry, initMethod, shortAlgo,
  longAlgo)
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

