% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,ClusterInit-method}
\alias{initialize,ClusterInit-method}
\title{Create an instance of the [\code{\linkS4class{ClusterInit}}] class using new/initialize.}
\usage{
\S4method{initialize}{ClusterInit}(.Object, method = "random", nbInit = 5,
  algo = clusterAlgo("SEM", 20, 0))
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

