% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,IClusterModel-method}
\alias{initialize,IClusterModel-method}
\title{Initialize an instance of the [\code{\linkS4class{IClusterModel}}] class.}
\usage{
\S4method{initialize}{IClusterModel}(.Object, data, nbCluster, modelName)
}
\arguments{
\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialization method. Used internally in the `rtkpp' package.
}
\keyword{internal}

