% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,ClusterAlgo-method}
\alias{initialize,ClusterAlgo-method}
\alias{initialize,ClusterCategorical-method}
\alias{initialize,ClusterDiagGaussian-method}
\alias{initialize,ClusterGamma-method}
\alias{initialize,ClusterInit-method}
\alias{initialize,ClusterStrategy-method}
\alias{initialize,IClusterModel-method}
\title{Initialize an instance of a rtkpp class.}
\usage{
\S4method{initialize}{ClusterAlgo}(.Object, algo, nbIteration, epsilon)

\S4method{initialize}{ClusterCategorical}(.Object, data, nbCluster = 2,
  modelName = "categorical_pk_pjk")

\S4method{initialize}{ClusterDiagGaussian}(.Object, data, nbCluster = 2,
  modelName = "gaussian_pk_sjk")

\S4method{initialize}{ClusterGamma}(.Object, data, nbCluster = 2,
  modelName = "gamma_pk_ajk_bjk")

\S4method{initialize}{ClusterInit}(.Object, method = "random",
  nbInitRun = 5, algo = clusterAlgo("SEM", 20, 0))

\S4method{initialize}{ClusterStrategy}(.Object, nbTry, nbShortRun, initMethod,
  shortAlgo, longAlgo)

\S4method{initialize}{IClusterModel}(.Object, data, nbCluster, modelName)
}
\description{
Initialization method of the [\code{\linkS4class{ClusterAlgo}}] class.
Used internally in the `rtkpp' package.

Initialization method of the [\code{\linkS4class{ClusterCategorical}}] class.
Used internally in the `rtkpp' package.

Initialization method of the [\code{\linkS4class{ClusterDiagGaussian}}] class.
Used internally in the `rtkpp' package.

Initialization method of the [\code{\linkS4class{ClusterGamma}}] class.
Used internally in the `rtkpp' package.

Initialization method of the [\code{\linkS4class{ClusterInit}}] class.
Used internally in the `rtkpp' package.

Initialization method of the [\code{\linkS4class{ClusterStrategy}}] class.
Used internally in the `rtkpp' package.

Initialization method of the [\code{\linkS4class{IClusterModel}}] class.
Used internally in the `rtkpp' package.
}
\keyword{internal}

