% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,ClusterDiagGaussian-method}
\alias{plot,ClusterDiagGaussian-method}
\alias{plot-ClusterDiagGaussian}
\title{Plotting of a class [\code{\linkS4class{ClusterDiagGaussian}}]}
\usage{
\S4method{plot}{ClusterDiagGaussian}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterDiagGaussian}}]}

\item{y}{a list of variables to plot (subset). Variables names or indices.
If missing all the variables are represented.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterDiagGaussian}}] object
using the estimated parameters and partition.
}
\examples{
## the famous iris data set
  data(iris)
  model <- clusterDiagGaussian(iris[1:4], 3, strategy = clusterFastStrategy())
\dontrun{
  plot(model)
  plot(model, c(1,3))
  plot(model, c("Sepal.Length","Sepal.Width"))
  }
}
\seealso{
\code{\link{plot}}
}

