% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,ClusterGamma-method}
\alias{ClusterGamma-method}
\alias{plot,ClusterGamma-method}
\alias{plot-ClusterGamma,}
\title{Plotting of a class [\code{\linkS4class{ClusterGamma}}]}
\usage{
\S4method{plot}{ClusterGamma}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterGamma}}]}

\item{y}{a list of variables to plot (subset). Variables names or indices.
If missingValues all the variables are represented.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterGamma}}] object
using the estimated parameters and partition.
}
\examples{
## for quantitative case
  data(iris)
  model <- clusterGamma(iris[1:4], 3, strategy = clusterFastStrategy())
\dontrun{
  plot(model)
  plot(model, c(1,3))
  plot(model, c("Sepal.Length","Sepal.Width"))
  }
}
\seealso{
\code{\link{plot}}
}

