% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,ClusterHeterogeneous-method}
\alias{plot,ClusterHeterogeneous-method}
\alias{plot-ClusterHeterogeneous}
\title{Plotting of a class [\code{\linkS4class{ClusterHeterogeneous}}]}
\usage{
\S4method{plot}{ClusterHeterogeneous}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterHeterogeneous}}]}

\item{y}{a number between 1 and K-1.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterHeterogeneous}}] object
using the estimated parameters and partition.
}
\examples{
\dontrun{
  ## the car data set
  data(car)
  model <- clusterHeterogeneous(car, 3, strategy = clusterFastStrategy())
  plot(model)
  }
}
\seealso{
\code{\link{plot}}
}

