% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,ClusterPoisson-method}
\alias{ClusterPoisson-method}
\alias{plot,ClusterPoisson-method}
\alias{plot-ClusterPoisson,}
\title{Plotting of a class [\code{\linkS4class{ClusterPoisson}}]}
\usage{
\S4method{plot}{ClusterPoisson}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterPoisson}}]}

\item{y}{a list of variables to plot (subset). Variables names or indices.
If missingValues all the variables are represented.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterPoisson}}] object
using the estimated parameters and partition.
}
\examples{
## for quantitative case
  data(DebTrivedi)
  dt <- DebTrivedi[, c(1, 6,8, 15)]
  model <- clusterPoisson(iris[1:4], 3, strategy = clusterFastStrategy())
\dontrun{
  plot(model)
  plot(model, c(1,2))
  }
}
\seealso{
\code{\link{plot}}
}

