% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,ClusterDiagGaussianModel-method}
\title{Plotting of a class [\code{\linkS4class{ClusterDiagGaussianModel}}]}
\usage{
\S4method{plot}{ClusterDiagGaussianModel}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterDiagGaussianModel}}]}

\item{y}{a list of variables to plot (subset). Variables names or indices.
If missing all the variables are represented.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterDiagGaussianModel}}] object
using the estimated parameters and partition.
}
\examples{
## for quantitative case
  data(iris)
  mixt <- clusterDiagGaussian(iris[1:4],3)
\dontrun{
  plot(mixt)
  plot(mixt, c(1,3))
  plot(mixt, c("Sepal.Length","Sepal.Width"))
  }
}
\seealso{
\code{\link{plot}}
}

