% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{print,ClusterAlgo-method}
\alias{print}
\alias{print,ClusterAlgo-method}
\alias{print,ClusterDiagGaussianModel-method}
\alias{print,ClusterInit-method}
\alias{print,ClusterStrategy-method}
\alias{print,IClusterModel-method}
\alias{print,MixmodCluster-method}
\alias{print-algo,ClusterAlgo-method}
\alias{print-init,ClusterInit,ClusterInit-method}
\alias{print-strategy,ClusterStrategy-method}
\title{Print a rtkpp class to standard output.}
\usage{
\S4method{print}{ClusterAlgo}(x, ...)

\S4method{print}{ClusterInit}(x, ...)

\S4method{print}{IClusterModel}(x, ...)

\S4method{print}{ClusterDiagGaussianModel}(x, ...)

\S4method{print}{ClusterStrategy}(x, ...)
}
\arguments{
\item{x}{a rtkpp object: a \code{\linkS4class{ClusterStrategy}},
a \code{\linkS4class{ClusterInit}} or a \code{\linkS4class{ClusterAlgo}}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Prints to standard out.
}
\description{
Print a rtkpp class to standard output.
}
\examples{
## for strategy
  strategy <- clusterStrategy()
  print(strategy)
}
\seealso{
\code{\link{print}}
}

