% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rtkpp}
\alias{rtkpp}
\alias{rtkpp-package}
\title{rtkpp is a R/stk++ bridge package based on the existing STK++ library.}
\description{
This package contains the header files for the STK++ template
library. The typical usage is to install this package and list it in
the \env{LinkingTo: } line in the \file{DESCRIPTION} file of
other packages.
}
\details{
As described at the STK++ project's home page, \url{http://stkpp.org},
STK++ is a versatile, fast, reliable and elegant collection of C++ classes
for statistics, clustering, linear algebra, regression, dimension reduction,
etc. Some functionalities provided by the library are available in the R
environment.

\tabular{ll}{
  Package: \tab rtkpp\cr
  Type: \tab Package\cr
  Version: \tab 2.0.1\cr
  Date: \tab 2014-07-05\cr
  License: \tab LGPL + file LICENSE\cr
  LazyLoad: \tab yes\cr
}
}
\author{
Author: Serge Iovleff \email{contact@stkpp.org}
}
\keyword{stk++}

