% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtkpp.system.file}
\alias{CxxFlags}
\alias{LdFlags}
\alias{asBuildPath}
\alias{rtkpp.system.file}
\alias{rtkppLdFlags}
\title{make sure system.file returns an absolute path}
\usage{
rtkpp.system.file(...)

rtkppLdFlags()

CxxFlags(cpp11 = FALSE)

LdFlags()

asBuildPath(path)
}
\description{
make sure system.file returns an absolute path

Provide linker flags -- i.e. /path/to/libSTKpp.a.

CxxFlags defaults

LdFlags defaults

Transform a path for passing to the build system on the command line.
Leave paths alone for posix. For Windows, mirror the behavior of the
R package build system by starting with the fully resolved absolute path,
transforming it to a short path name if it contains spaces, and then
converting backslashes to forward slashes
}
\keyword{internal}

