% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,ClusterCategorical-method}
\alias{summary}
\alias{summary,ClusterCategorical-method}
\alias{summary,ClusterDiagGaussian-method}
\alias{summary,ClusterGamma-method}
\alias{summary,IClusterModel-method}
\alias{summary,MixmodResults-method}
\title{Produce summary of a rtkpp class.}
\usage{
\S4method{summary}{ClusterCategorical}(object, ...)

\S4method{summary}{ClusterDiagGaussian}(object, ...)

\S4method{summary}{ClusterGamma}(object, ...)

\S4method{summary}{IClusterModel}(object, ...)
}
\arguments{
\item{object}{any cluster model deriving from a \code{\linkS4class{IClusterModel}} object.}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Summaries to standard out.
}
\description{
Produce summary of a rtkpp class.
}
\examples{
data(geyser)
  model <- clusterDiagGaussian(geyser,3)
  summary(model)
}
\seealso{
\code{\link{summary}}
}

