% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clusterDiagGaussian}
\alias{clusterDiagGaussian}
\title{Create an instance of the [\code{\linkS4class{ClusterDiagGaussianModel}}] class}
\usage{
clusterDiagGaussian(data, nbCluster = 2, modelNames = NULL,
  strategy = clusterStrategy(), criterion = "BIC")
}
\arguments{
\item{data}{frame or matrix containing the data. Rows correspond to observations and columns correspond to variables.}

\item{nbCluster}{numeric listing the number of clusters.}

\item{modelNames}{a list of models to run. By default all diagoanl Gaussian models are estimated.}

\item{strategy}{a [\code{\linkS4class{ClusterStrategy}}] object containing the strategy to run.
clusterStrategy() method by default.}

\item{criterion}{character defining the criterion to select the best model. The best model is the one with
the lowest criterion value. Possible values: "BIC", "AIC". Default is "BIC".}
}
\value{
An instance of the [\code{\linkS4class{ClusterDiagGaussianModel}}] class.
}
\description{
This function computes an optimal diagonal Gaussian mixture model according
to the criteria furnished and the list of model defined in
[\code{modelNames}], using the
algorithms specified in [\code{modelNames}].
}
\examples{
## A quantitative example with the famous geyser data set
  data(geyser)
  ## with default values
  clusterDiagGaussian(geyser, nbCluster=2:6)
  ## use graphics functions
  model <- clusterDiagGaussian(data=geyser, nbCluster=2:3)
  \dontrun{
  plot(model)
  }
  ## print model
  print(model)
  ## get summary
  summary(model)
}
\author{
Serge Iovleff
}

