% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diagGaussianNames}
\alias{diagGaussianNames}
\alias{validDiagGaussianNames}
\alias{validGammaNames}
\title{diagGaussianNames: Build a vector of Gaussian model names.

In a Gaussian model, we assume that the variance
matrices are diagonal in each cluster. This gives rise to 8 models:
\enumerate{
 \item {The proportions can be equal or free.}
 \item {The standard deviations can be equal or free for all the variables.}
 \item {The standard deviations can be equal or free for all the clusters.}
}

The model names are summarized in the following array:
\tabular{llll}{
 Model Name      \tab Proportions \tab Variances in variables \tab Variances in clusters \cr
 gaussian_p_sjk  \tab Equal \tab Free                  \tab Free  \cr
 gaussian_p_sj   \tab Equal \tab Free                  \tab Equal \cr
 gaussian_p_sk   \tab Equal \tab Equal                 \tab Free  \cr
 gaussian_p_s    \tab Equal \tab Equal                \tab Equal \cr
 gaussian_pk_sjk \tab Free  \tab Free                 \tab Free  \cr
 gaussian_pk_sj  \tab Free  \tab Free                 \tab Equal \cr
 gaussian_pk_sk  \tab Free  \tab Equal                \tab Free  \cr
 gaussian_pk_s   \tab Free  \tab Equal                \tab Equal \cr
}}
\usage{
diagGaussianNames(prop = "all", sdInCluster = "all",
  sdBetweenCluster = "all")

validDiagGaussianNames(names)

validGammaNames(names)
}
\arguments{
\item{prop}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{sdInCluster}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{sdBetweenCluster}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{names}{a vector of character}

\item{names}{a vector of character}
}
\value{
A vector of character with the models names.
}
\description{
diagGaussianNames: Build a vector of Gaussian model names.

In a Gaussian model, we assume that the variance
matrices are diagonal in each cluster. This gives rise to 8 models:
\enumerate{
 \item {The proportions can be equal or free.}
 \item {The standard deviations can be equal or free for all the variables.}
 \item {The standard deviations can be equal or free for all the clusters.}
}

The model names are summarized in the following array:
\tabular{llll}{
 Model Name      \tab Proportions \tab Variances in variables \tab Variances in clusters \cr
 gaussian_p_sjk  \tab Equal \tab Free                  \tab Free  \cr
 gaussian_p_sj   \tab Equal \tab Free                  \tab Equal \cr
 gaussian_p_sk   \tab Equal \tab Equal                 \tab Free  \cr
 gaussian_p_s    \tab Equal \tab Equal                \tab Equal \cr
 gaussian_pk_sjk \tab Free  \tab Free                 \tab Free  \cr
 gaussian_pk_sj  \tab Free  \tab Free                 \tab Equal \cr
 gaussian_pk_sk  \tab Free  \tab Equal                \tab Free  \cr
 gaussian_pk_s   \tab Free  \tab Equal                \tab Equal \cr
}

check if a vector of diagonal Gaussian model name is correct.

check if a vector of gamma model name is correct.
}
\examples{
diagGaussianNames()
diagGaussianNames("all", "equal", "free") # same as c("gaussian_p_sk", "gaussian_pk_sk")
}
\keyword{internal}

