% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{print,ClusterAlgo-method}
\alias{print}
\alias{print,ClusterAlgo-method}
\alias{print,ClusterCategorical-method}
\alias{print,ClusterCategoricalComponent-method}
\alias{print,ClusterDiagGaussian-method}
\alias{print,ClusterDiagGaussianComponent-method}
\alias{print,ClusterGamma-method}
\alias{print,ClusterGammaComponent-method}
\alias{print,ClusterHeterogeneous-method}
\alias{print,ClusterInit-method}
\alias{print,ClusterPoisson-method}
\alias{print,ClusterPoissonComponent-method}
\alias{print,ClusterStrategy-method}
\alias{print,IClusterComponent-method}
\alias{print,IClusterModelBase-method}
\alias{print-algo,ClusterAlgo-method}
\alias{print-init,ClusterInit,ClusterInit-method}
\alias{print-strategy,ClusterStrategy-method}
\title{Print a rtkpp class to standard output.}
\usage{
\S4method{print}{ClusterAlgo}(x, ...)

\S4method{print}{ClusterCategoricalComponent}(x, k, ...)

\S4method{print}{ClusterCategorical}(x, ...)

\S4method{print}{ClusterDiagGaussianComponent}(x, k, ...)

\S4method{print}{ClusterDiagGaussian}(x, ...)

\S4method{print}{ClusterGammaComponent}(x, k, ...)

\S4method{print}{ClusterGamma}(x, ...)

\S4method{print}{ClusterHeterogeneous}(x, ...)

\S4method{print}{ClusterInit}(x, ...)

\S4method{print}{ClusterPoissonComponent}(x, k, ...)

\S4method{print}{ClusterPoisson}(x, ...)

\S4method{print}{ClusterStrategy}(x, ...)

\S4method{print}{IClusterModelBase}(x, ...)

\S4method{print}{IClusterComponent}(x, ...)
}
\arguments{
\item{x}{a rtkpp object: a \code{\linkS4class{ClusterStrategy}},
a \code{\linkS4class{ClusterInit}} or a \code{\linkS4class{ClusterAlgo}}.}

\item{...}{further arguments passed to or from other methods}

\item{k}{the number of the cluster to print}
}
\value{
NULL. Prints to standard out.
}
\description{
Print a rtkpp class to standard output.
}
\examples{
## for cluster strategy
  strategy <- clusterStrategy()
  print(strategy)
  ## for cluster init
  init <- clusterInit()
  print(init)
  ## for cluster algo
  algo <- clusterAlgo()
  print(algo)
}
\seealso{
\code{\link{print}}
}

