% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CxxFlags}
\alias{CppFlags}
\alias{CxxFlags}
\alias{LdFlags}
\title{CxxFlags defaults for the rtkpp}
\usage{
CxxFlags(cpp11 = FALSE)

CppFlags()

LdFlags()
}
\description{
CxxFlags defaults for the rtkpp

CppFlags defaults for the rtkpp

LdFlags defaults for the rtkpp package
}
\keyword{internal}

