% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,ClusterAlgo-method}
\alias{show}
\alias{show,ClusterAlgo-method}
\alias{show,ClusterCategorical-method}
\alias{show,ClusterDiagGaussian-method}
\alias{show,ClusterGamma-method}
\alias{show,ClusterInit-method}
\alias{show,ClusterStrategy-method}
\alias{show,IClusterModel-method}
\alias{show-ClusterCategorical,ClusterCategorical,ClusterCategorical-method}
\alias{show-ClusterDiagGaussian,ClusterDiagGaussian,ClusterDiagGaussian-method}
\alias{show-ClusterGamma,ClusterGamma,ClusterGamma-method}
\alias{show-algo,ClusterAlgo-method}
\alias{show-init,ClusterInit,ClusterInit-method}
\alias{show-strategy,ClusterStrategy-method}
\title{Show description of a rtkpp class to standard output.}
\usage{
\S4method{show}{ClusterAlgo}(object)

\S4method{show}{ClusterCategorical}(object)

\S4method{show}{ClusterDiagGaussian}(object)

\S4method{show}{ClusterGamma}(object)

\S4method{show}{ClusterInit}(object)

\S4method{show}{ClusterStrategy}(object)

\S4method{show}{IClusterModel}(object)
}
\arguments{
\item{object}{a rtkpp object: a \code{\linkS4class{ClusterStrategy}},
a \code{\linkS4class{ClusterInit}} or a \code{\linkS4class{ClusterAlgo}}.}
}
\value{
NULL. Prints to standard out.
}
\description{
Show description of a rtkpp class to standard output.
}
\examples{
## for strategy
  strategy <- clusterStrategy()
  show(strategy)
  ## for cluster init
  init <- clusterInit()
  show(init)
  ## for cluster algo
  algo <- clusterAlgo()
  show(algo)
}
\seealso{
\code{\link{show}}
}

