% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,ClusterAlgo-method}
\alias{show}
\alias{show,ClusterAlgo-method}
\alias{show,ClusterCategorical-method}
\alias{show,ClusterCategoricalComponent-method}
\alias{show,ClusterDiagGaussian-method}
\alias{show,ClusterDiagGaussianComponent-method}
\alias{show,ClusterGamma-method}
\alias{show,ClusterGammaComponent-method}
\alias{show,ClusterHeterogeneous-method}
\alias{show,ClusterInit-method}
\alias{show,ClusterPoisson-method}
\alias{show,ClusterPoissonComponent-method}
\alias{show,ClusterStrategy-method}
\alias{show,IClusterComponent-method}
\alias{show,IClusterModelBase-method}
\alias{show-ClusterCategorical,ClusterCategorical,ClusterCategorical-method}
\alias{show-ClusterCategoricalComponent,ClusterCategoricalComponent,ClusterCategoricalComponent-method}
\alias{show-ClusterDiagGaussian,ClusterDiagGaussian,ClusterDiagGaussian-method}
\alias{show-ClusterDiagGaussianComponent,ClusterDiagGaussianComponent,ClusterDiagGaussianComponent-method}
\alias{show-ClusterGamma,ClusterGamma,ClusterGamma-method}
\alias{show-ClusterGammaComponent,ClusterGammaComponent,ClusterGammaComponent-method}
\alias{show-ClusterHeterogeneous,ClusterHeterogeneous,ClusterHeterogeneous-method}
\alias{show-ClusterPoisson,ClusterPoisson,ClusterPoisson-method}
\alias{show-ClusterPoissonComponent,ClusterPoissonComponent,ClusterPoissonComponent-method}
\alias{show-IClusterComponent,IClusterComponent,IClusterComponent-method}
\alias{show-algo,ClusterAlgo-method}
\alias{show-init,ClusterInit,ClusterInit-method}
\alias{show-strategy,ClusterStrategy-method}
\title{Show description of a rtkpp class to standard output.}
\usage{
\S4method{show}{ClusterAlgo}(object)

\S4method{show}{ClusterCategoricalComponent}(object)

\S4method{show}{ClusterCategorical}(object)

\S4method{show}{ClusterDiagGaussianComponent}(object)

\S4method{show}{ClusterDiagGaussian}(object)

\S4method{show}{ClusterGammaComponent}(object)

\S4method{show}{ClusterGamma}(object)

\S4method{show}{ClusterHeterogeneous}(object)

\S4method{show}{ClusterInit}(object)

\S4method{show}{ClusterPoissonComponent}(object)

\S4method{show}{ClusterPoisson}(object)

\S4method{show}{ClusterStrategy}(object)

\S4method{show}{IClusterModelBase}(object)

\S4method{show}{IClusterComponent}(object)
}
\arguments{
\item{object}{a rtkpp object: a \code{\linkS4class{ClusterStrategy}},
a \code{\linkS4class{ClusterInit}} or a \code{\linkS4class{ClusterAlgo}}.}
}
\value{
NULL. Prints to standard out.
}
\description{
Show description of a rtkpp class to standard output.
}
\examples{
## for strategy
  strategy <- clusterStrategy()
  show(strategy)
  ## for cluster init
  init <- clusterInit()
  show(init)
  ## for cluster algo
  algo <- clusterAlgo()
  show(algo)
}
\seealso{
\code{\link{show}}
}

